Temp fix for damage invincibility
For some reason, dashing into a pile of lava will instantly kill you, because the damage triggers on you a bunch of times.

I think that the issue might be that damage doesn't properly check if the player's damage collision is on, and somehow deals damage anyways. I don't know exactly how collision works in Unity yet, but this should only be a temp fix, which is manually checking if damage collision should be active.